Resources
Added the excel file with the list of Admin roles


Roles Verification

Was able to look up roles one entry at a time.
Error occurs at permission("View a user's roles") with single quote (')
Changed the csvFile to xlsxFile and directed to the local directory of
the added excel file of Admin roles.